@@ -1,22 +1,32 @@
 {
-  "name": "caffeine-counter",
+  "name": "caffeine-tracker",
   "version": "1.0.0",
-  "description": "Website to track caffeine intake",
+  "description": "",
   "main": "server.js",
   "scripts": {
     "start": "node server.js",
-    "dev": "nodemon server.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "dev": "nodemon server.js"
   },
-  "author": "FP",
   "dependencies": {
+    "chalk": "^5.6.0",
+    "connect-mongo": "^5.1.0",
     "express": "^4.18.2",
-    "mongoose": "^7.5.0"
+    "express-session": "^1.17.3",
+    "google-auth-library": "^9.4.1",
+    "mongoose": "^7.6.3"
   },
   "devDependencies": {
     "nodemon": "^3.0.1"
   },
   "engines": {
-    "node": ">=14.0.0"
-  }
+    "node": ">=16.0.0"
+  },
+  "keywords": [
+    "caffeine",
+    "tracker",
+    "social",
+    "authentication",
+    "leaderboard",
+    "mongodb"
+  ]
 }
